Proyecto1_Fase2/Apis/Flask-Api/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0